Improve Forst Create Bonus metod, remove some await, works
